Makefile:  this is just a quick makefile that I threw together.  This should be sufficient for the project, though I might want to add other targets for tests of various vigors.  For example, "make test-running-time" would produce graphs.
fibheap.[ch]: This is the initial version of the heap that I'm adding.  It
requires a *lot* of cleaning up and commenting to be presentable.  There's
probably a bug or two lying around as well.  Need to figure out how to deal with
instrumentation (turning it on and off).  Also, need to figure out how to pull
the constant down just a bit on extract-min.  Seems like c=25 could be reduced.

test-fib.c:  This is merely a stub of the test suite.  I alreadly have a very
rough, nasty test file in another repository that I'll co-opt.  But, I need to
find if first ;-).